requirements.txt
gunicorn
dash_bootstrap_components
djangorestframework
django_plotly_dash
dj-database-url
whitenoise
pillow
dajngo